+++ app/domain/authentication/installed_authenticators.rb
@@ -3,15 +3,44 @@
 module Authentication
   class InstalledAuthenticators
 
-    def self.new(env, authentication_module: ::Authentication)
-      ::Util::Submodules.of(authentication_module)
-        .flat_map { |mod| ::Util::Submodules.of(mod) }
+    AUTHN_RESOURCE_PREFIX = "conjur/authn-"
+
+    def self.authenticators(env, authentication_module: ::Authentication)
+      loaded_authenticators(authentication_module)
         .select { |cls| valid?(cls) }
         .map { |cls| [url_for(cls), authenticator_instance(cls, env)] }
         .to_h
     end
 
+    def self.login_authenticators(env, authentication_module: ::Authentication)
+      loaded_authenticators(authentication_module)
+        .select { |cls| provides_login?(cls) }
+        .map { |cls| [url_for(cls), authenticator_instance(cls, env)] }
+        .to_h
+    end
+
+    def self.configured_authenticators
+      identifier = Sequel.function(:identifier, :resource_id)
+      kind = Sequel.function(:kind, :resource_id)
+
+      Resource
+        .where(identifier.like("#{AUTHN_RESOURCE_PREFIX}%"))
+        .where(kind => "webservice")
+        .select_map(identifier)
+        .map { |id| id.sub %r{^conjur\/}, "" }
+        .push(::Authentication::Strategy.default_authenticator_name)
+    end
+
+    def self.enabled_authenticators(env)
+      (env["CONJUR_AUTHENTICATORS"] || ::Authentication::Strategy.default_authenticator_name).split(",")
+    end
+
     private
+
+    def self.loaded_authenticators(authentication_module)
+      ::Util::Submodules.of(authentication_module)
+        .flat_map { |mod| ::Util::Submodules.of(mod) }
+    end
 
     def self.authenticator_instance(cls, env)
       pass_env = ::Authentication::AuthenticatorClass.new(cls).requires_env_arg?
@@ -26,5 +55,9 @@ module Authentication
       ::Authentication::AuthenticatorClass::Validation.new(cls).valid?
     end
 
+    def self.provides_login?(cls)
+      validation = ::Authentication::AuthenticatorClass::Validation.new(cls)
+      validation.valid? && validation.provides_login?
+    end
   end
 end