+++ cucumber/authenticators/features/step_definitions/authn_ldap_steps.rb
@@ -6,24 +6,30 @@ Given(/^a policy:$/) do |policy|
   load_root_policy(policy)
 end
 
+When(/I login via LDAP as (?:\S)+ Conjur user "(\S+)"/) do |username|
+  login_with_ldap(service_id: 'test', account: 'cucumber', 
+                  username: username, password: username)
+end
+
 # First non-captured group allows for adjectives to clarify the
 # purpose of the test.  They aren't actually used.
 #
-When(/I login via LDAP as (?:\S)+ Conjur user "(\S+)"/) do |username|
+When(/I authenticate via LDAP as (?:\S)+ Conjur user "(\S+)"( using key)?/) do |username, using_key|
+  password = using_key ? ldap_auth_key : username
   authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
-                         username: username, password: username)
+                         username: username, api_key: password)
 end
 
 When(/my LDAP password for (?:\S)+ Conjur user "(\S+)" is empty/) do |username|
-  authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
-                         username: username, password: '')
+  login_with_ldap(service_id: 'test', account: 'cucumber', 
+                  username: username, password: '')
 end
 
 
 
 When(/my LDAP password is wrong for authorized user "(\S+)"/) do |username|
-  authenticate_with_ldap(service_id: 'test', account: 'cucumber', 
-                         username: username, password: 'BAD_PASSWORD')
+  login_with_ldap(service_id: 'test', account: 'cucumber', 
+                  username: username, password: 'BAD_PASSWORD')
 end
 
 Then(/"(\S+)" is authorized/) do |username|
@@ -31,5 +37,9 @@ Then(/"(\S+)" is authorized/) do |username|
 end
 
 Then(/it is denied/) do
-  expect(authorized?).to be true
+  expect(unauthorized?).to be true
+end
+
+Then(/it is forbidden/) do
+  expect(forbidden?).to be true
 end