+++ spec/controllers/credentials_controller_spec.rb
@@ -5,49 +5,12 @@ require 'spec_helper'
 describe CredentialsController, :type => :controller do
   let(:login) { "u-#{random_hex}" }
   let(:account) { "rspec" }
-
-  context "#login" do
-    shared_examples_for "successful authentication" do
-      it "succeeds" do
-        post :login, params
-        expect(response).to be_ok
-        expect(response.body.length).to be >= 44
-      end
-    end
-    
-    shared_examples_for "authentication denied" do
-      it "is unauthorized" do
-        post :login, account: account
-        expect(response.code).to eq("401")
-      end
-    end
-    
-    context "without auth" do
-      it_should_behave_like "authentication denied"
-    end
-    context "when user doesn't exist" do
-      let(:basic_password) { "the-password" }
-      include_context "authenticate Basic"
-      it_should_behave_like "authentication denied"
-    end
-    context "when user exists" do
-      include_context "create user"
-      context "with basic auth" do
-        let(:basic_password) { api_key }
-        include_context "authenticate Basic"
-        it_should_behave_like "successful authentication"
-      end
-      context "with Token auth" do
-        include_context "authenticate Token"
-        it_should_behave_like "authentication denied"
-      end
-    end
-  end
+  let(:authenticator) { "authn" }
   
   context "#rotate_api_key" do
     shared_examples_for "authentication denied" do
       it "is unauthorized" do
-        post :rotate_api_key, account: account
+        post :rotate_api_key, account: account, authenticator: authenticator
         expect(response.code).to eq("401")
       end
     end
@@ -71,7 +34,7 @@ describe CredentialsController, :type => :controller do
         include_context "create user"
         it "rotates the user's API key" do
           api_key = the_user.api_key
-          post :rotate_api_key, account: account
+          post :rotate_api_key, account: account, authenticator: authenticator
           the_user.credentials.reload
           expect(response.code).to eq("200")
           expect(response.body).to eq(the_user.credentials.api_key)
@@ -85,7 +48,7 @@ describe CredentialsController, :type => :controller do
     let(:new_password) { +"new-password" }
     context "without auth" do
       it "is unauthorized" do
-        post :update_password, account: account
+        post :update_password, account: account, authenticator: authenticator
         expect(response.code).to eq("401")
       end
     end
@@ -110,7 +73,7 @@ describe CredentialsController, :type => :controller do
           }
         }
         it "is is malformed" do
-          post :update_password, account: account
+          post :update_password, account: account, authenticator: authenticator
           expect(response.code).to eq("422")
           expect(response.body).to eq(errors.to_json)
         end
@@ -119,7 +82,7 @@ describe CredentialsController, :type => :controller do
         before { request.env['RAW_POST_DATA'] = new_password }
         context "and valid password" do
           it "updates the password" do
-            post :update_password, account: account
+            post :update_password, account: account, authenticator: authenticator
             expect(response.code).to eq("204")
             the_user.credentials.reload
             expect(the_user.credentials.authenticate(new_password)).to be(true)
@@ -143,7 +106,7 @@ describe CredentialsController, :type => :controller do
             }
           }
           it "reports the error" do
-            post :update_password, account: account
+            post :update_password, account: account, authenticator: authenticator
             expect(response.code).to eq("422")
             expect(response.body).to eq(errors.to_json)
             expect(the_user.credentials.authenticate(new_password)).to be(false)